+++ src/ssd/ssd_border.c
@@ -2,7 +2,7 @@
 
 #include "common/scene-helpers.h"
 #include "labwc.h"
-#include "ssd.h"
+#include "ssd-internal.h"
 #include "theme.h"
 #include "view.h"
