+++ src/ssd/ssd_titlebar.c
@@ -9,7 +9,7 @@
 #include "common/scene-helpers.h"
 #include "labwc.h"
 #include "node.h"
-#include "ssd.h"
+#include "ssd-internal.h"
 #include "theme.h"
 #include "view.h"
